revert demo changes
